Configure ProvisionedConcurrentExecutions
https://docs.aws.amazon.com/serverless-application-model/latest/developerguide/sam-resource-function.html#sam-function-provisionedconcurrencyconfig
(link shows information about ProvisionedConcurrencyConfig and
ReservedConcurrentExecutions

This describes how to configure the ProvisionedConcurrencyConfig :
https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-lambda-alias-provisionedconcurrencyconfiguration.html

and ReservedConcurrentExecutions is:
https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-lambda-function.html#cfn-lambda-function-reservedconcurrentexecutions

From what I can understand:

ReservedConcurrentExecutions: The number of simultaneous executions to
reserve for the function.

ProvisionedConcurrentExecutions (under
ProvisionedConcurrencyConfiguration ): The amount of provisioned
concurrency to allocate for the alias.

I believe we actually want to configure ProvisionedConcurrentExecutions
, as of now we don't have any need for configure a maximum concurrency
as well.